# [ENH] Speed up spatial normalization