--- a/mriqc/data/itk_identity.tfm
+++ b/mriqc/data/itk_identity.tfm
@@ -1,0 +1,5 @@
+#Insight Transform File V1.0
+#Transform 0
+Transform: IdentityTransform_double_3_3
+Parameters: 
+FixedParameters: