--- a/mriqc/interfaces/common.py
+++ b/mriqc/interfaces/common.py
@@ -165,14 +165,18 @@ class EnsureSize(SimpleInterface):
     def _run_interface(self, runtime):
         nii = nb.load(self.inputs.in_file)
         zooms = nii.header.get_zooms()
-        if np.all(np.array(zooms[:3]) >= self.inputs.pixel_size):
+        size_diff = np.array(zooms[:3]) - self.inputs.pixel_size
+        if np.all(size_diff >= 0.1):
+            IFLOGGER.info('Voxel size is large enough')
             self._results['out_file'] = self.inputs.in_file
             if isdefined(self.inputs.in_mask):
                 self._results['out_mask'] = self.inputs.in_mask
             return runtime
 
-        IFLOGGER.info('One or more voxel dimensions (%f, %f, %f) are greater than the requested'
-                      'voxel size (%f)', zooms[0], zooms[1], zooms[2], self.inputs.pixel_size)
+        IFLOGGER.info('One or more voxel dimensions (%f, %f, %f) are smaller than '
+                      'the requested voxel size (%f) - diff=(%f, %f, %f)', zooms[0],
+                      zooms[1], zooms[2], self.inputs.pixel_size, size_diff[0],
+                      size_diff[1], size_diff[2])
 
         # Figure out new matrix
         # 1) Get base affine