--- a/mriqc/workflows/anatomical.py
+++ b/mriqc/workflows/anatomical.py
@@ -94,17 +94,7 @@ def anat_qc_workflow(dataset, settings, mod='T1w', name='anatMRIQC'):
     # 3. Head mask
     hmsk = headmsk_wf()
     # 4. Spatial Normalization, using ANTs
-    norm = pe.Node(RobustMNINormalization(
-        num_threads=settings.get('ants_nthreads'), template='mni_icbm152_nlin_asym_09c',
-        testing=settings.get('testing', False), generate_report=True), name='SpatialNormalization')
-    norm.interface.num_threads = settings.get('ants_nthreads')
-    norm.interface.estimated_memory_gb = 6
-
-    if mod == 'T1w':
-        norm.inputs.reference = 'T1'
-    elif mod == 'T2w':
-        norm.inputs.reference = 'T2'
-
+    norm = spatial_normalization(settings)
     # 5. Air mask (with and without artifacts)
     amw = airmsk_wf()
     # 6. Brain tissue segmentation
@@ -132,12 +122,15 @@ def anat_qc_workflow(dataset, settings, mod='T1w', name='anatMRIQC'):
         (asw, segment, [('outputnode.out_file', 'in_files')]),
         (asw, hmsk, [('outputnode.bias_corrected', 'inputnode.in_file')]),
         (segment, hmsk, [('tissue_class_map', 'inputnode.in_segm')]),
-        (asw, norm, [('outputnode.bias_corrected', 'moving_image'),
-                     ('outputnode.out_mask', 'moving_mask')]),
+        (asw, norm, [('outputnode.bias_corrected', 'inputnode.moving_image'),
+                     ('outputnode.out_mask', 'inputnode.moving_mask')]),
+        (norm, amw, [
+            ('outputnode.inverse_composite_transform', 'inputnode.inverse_composite_transform')]),
+        (norm, iqmswf, [
+            ('outputnode.inverse_composite_transform', 'inputnode.inverse_composite_transform')]),
+        (norm, repwf, ([
+            ('outputnode.out_report', 'inputnode.mni_report')])),
         (to_ras, amw, [('out_file', 'inputnode.in_file')]),
-        (norm, amw, [('inverse_composite_transform', 'inputnode.inverse_composite_transform')]),
-        (norm, iqmswf, [('inverse_composite_transform', 'inputnode.inverse_composite_transform')]),
-        (norm, repwf, ([('out_report', 'inputnode.mni_report')])),
         (asw, amw, [('outputnode.out_mask', 'inputnode.in_mask')]),
         (hmsk, amw, [('outputnode.out_file', 'inputnode.head_mask')]),
         (to_ras, iqmswf, [('out_file', 'inputnode.orig')]),
@@ -162,6 +155,57 @@ def anat_qc_workflow(dataset, settings, mod='T1w', name='anatMRIQC'):
         (iqmswf, outputnode, [('outputnode.out_file', 'out_json')])
     ])
 
+    return workflow
+
+def spatial_normalization(settings, mod='T1w', name='SpatialNormalization'):
+    """
+    A simple workflow to perform spatial normalization
+
+    """
+    from nipype.interfaces.ants import AffineInitializer
+    from niworkflows.data import getters as niwgetters
+
+    # Have some settings handy
+    test_mode = settings.get('testing', False)
+    tpl_id = settings.get('template_id', 'mni_icbm152_nlin_asym_09c')
+    mni_template = getattr(niwgetters, 'get_{}'.format(tpl_id))()
+    res = 1 if not test_mode else 2
+
+
+    # Define workflow interface
+    workflow = pe.Workflow(name=name)
+    inputnode = pe.Node(niu.IdentityInterface(fields=[
+        'moving_image', 'moving_mask']), name='inputnode')
+    outputnode = pe.Node(niu.IdentityInterface(fields=[
+        'inverse_composite_transform', 'out_report']), name='outputnode')
+
+    # Mask inputs for initialization
+    mmask = pe.Node(fsl.ApplyMask(), name='MovingApplyMask')
+    fmask = pe.Node(fsl.ApplyMask(), name='FixedApplyMask')
+    fmask.inputs.in_file = op.join(mni_template, '%dmm_%s.nii.gz' % (res, mod[:2]))
+    fmask.inputs.mask_file = op.join(mni_template, '%dmm_brainmask.nii.gz' % res)
+
+    # Initializer
+    init = pe.Node(AffineInitializer(), name='NormalizationInit')
+
+    # Spatial normalization
+    norm = pe.Node(RobustMNINormalization(
+        num_threads=settings.get('ants_nthreads'), template=tpl_id, testing=True,
+        generate_report=True, reference=mod[:2]), name='SpatialNormalization')
+    norm.interface.num_threads = settings.get('ants_nthreads')
+    norm.interface.estimated_memory_gb = 6 if not test_mode else 3
+
+    workflow.connect([
+        (inputnode, mmask, [('moving_image', 'in_file'),
+                            ('moving_mask', 'mask_file')]),
+        (mmask, init, [('out_file', 'moving_image')]),
+        (fmask, init, [('out_file', 'fixed_image')]),
+        (init, norm, [('out_file', 'initial_moving_transform')]),
+        (inputnode, norm, [('moving_image', 'moving_image'),
+                           ('moving_mask', 'moving_mask')]),
+        (norm, outputnode, [('inverse_composite_transform', 'inverse_composite_transform'),
+                            ('out_report', 'out_report')]),
+    ])
     return workflow
 
 def compute_iqms(settings, modality='T1w', name='ComputeIQMs'):