--- a/mriqc/workflows/functional.py
+++ b/mriqc/workflows/functional.py
@@ -698,7 +698,8 @@ def epi_mni_align(name='SpatialNormalization', ants_nthreads=6, testing=False, r
 
     norm = pe.Node(RobustMNINormalization(
         num_threads=ants_nthreads, template='mni_icbm152_nlin_asym_09c',
-        testing=testing, moving='EPI', generate_report=True),
+        flavor='testing' if testing else 'precise',
+        moving='EPI', generate_report=True),
                    name='EPI2MNI')
     norm.inputs.reference_image = pkgrf(
         'mriqc', 'data/mni/%dmm_T2_brain.nii.gz' % resolution)