--- a/setup.py
+++ b/setup.py
@@ -54,6 +54,7 @@ def main():
                                           'participants=mriqc.bin.subject_wrangler:main']},
         packages=find_packages(exclude=['*.tests']),
         package_data={'mriqc': ['data/*.yml',
+                                'data/*.tfm',
                                 'data/csv/*.csv',
                                 'data/rfc-nzs-abide-1.0.pklz',
                                 'data/rfc-nzs-full-1.0.pklz',