… CHANGES.txt
… circle.yml
… requirements.txt