Please note I've created two new labels: RTM and WIP. I will not mark them in the title anymore. In a future (impossible to tell when), I will write a bit of code to automatically generate the CHANGES file using `git log --merges`.

For this reason, titles of PRs must start with [ENH], [FIX] or [DOC] and have a CHANGES-ready text.

UPDATE: (cc/ @chrisfilo) some benchmarks of these new settings on my desktop (8cpus) and CircleCI. This is on fully sampled datasets (the new SpatialNormalization will subsample them to 2.0mm isotropic size):

| ITK_GLOBAL_DEFAULT_NUMBER_OF_THREADS | dendrite | circleCI |
|--------------------------------------|----------|----------|
| 1                                    | 127.20s  | 241.87s  |
| 2                                    | 70.10s   | 259.70s  |
| 3                                    | 51.10s   | 169.10s  |
| 4                                    | 43.30s   | 163.20s  |
| 6                                    | 42.90s   | 163.10s  |
| 8                                    | 37.10s   | 165.90s  |
| 16                                   | 41.20s   | 168.30s |

Right now, the typical runtime for the affine-only registration with highly decimated images is 180s.