@oesteban 2017-05-02
The command I'm running in CircleCI:

```
docker run -v ~/antsdata/:/ants -w /ants --rm=false -it -e ITK_GLOBAL_DEFAULT_NUMBER_OF_THREADS=3 --entrypoint="antsRegistration" poldracklab/mriqc:latest antsRegistration --collapse-output-transforms 1 --dimensionality 3 --float 1 --initial-moving-transform [ transform.mat, 0 ] --initialize-transforms-per-stage 0 --interpolation LanczosWindowedSinc --output [ ants_t1_to_mni, ants_t1_to_mni_Warped.nii.gz ] --transform Rigid[ 0.01 ] --metric Mattes[ fixed_masked.nii.gz, moving_masked.nii.gz, 1, 32, Random, 0.15 ] --convergence [ 1000, 1e-06, 20 ] --smoothing-sigmas 4.0vox --shrink-factors 4 --use-estimate-learning-rate-once 1 --use-histogram-matching 1 --transform Affine[ 0.08 ] --metric Mattes[ fixed_masked.nii.gz, moving_masked.nii.gz, 1, 32, Regular, 0.15 ] --convergence [ 500x250x100, 1e-06, 20 ] --smoothing-sigmas 4.0x2.0x0.0vox --shrink-factors 4x2x1 --use-estimate-learning-rate-once 1 --use-histogram-matching 1 --transform SyN[ 0.1, 3.0, 0.0 ] --metric Mattes[ fixed_masked.nii.gz, moving_masked.nii.gz, 1, 56, Regular, 0.25 ] --convergence [ 50x20, 1e-06, 10 ] --smoothing-sigmas 1.0x0.0vox --shrink-factors 2x1 --use-estimate-learning-rate-once 1 --use-histogram-matching 1 --winsorize-image-intensities [ 0.0, 1.0 ]  --write-composite-transform 1
```