@oesteban 2017-05-05
(closes #450)